Cleanup, fix deprecated pods paths
